+++ KermesseApp/Models/vw_productos.cs
@@ -17,7 +17,7 @@ namespace KermesseApp.Models
         public string nombre { get; set; }
         public string desc_presentacion { get; set; }
         public int cantidad { get; set; }
-        public decimal precio_venta { get; set; }
+        public double precio_venta { get; set; }
         public string comunidad { get; set; }
         public string catprod { get; set; }
         public int id_producto { get; set; }
